fixes, fixtures and test coverage for agent block_devices.py
