--- chroma-agent/chroma_agent/device_plugins/linux_components/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) 2017 Intel Corporation. All rights reserved.
-# Use of this source code is governed by a MIT-style
-# license that can be found in the LICENSE file.
